+++ bin/ann_flutter_flavor.dart
@@ -1,6 +1,7 @@
 import 'package:args/command_runner.dart';
 import '../lib_cli/commands/sync_command.dart';
 import '../lib_cli/commands/validate_command.dart';
+import '../lib_cli/commands/validate_testspec_command.dart';
 import '../lib_cli/commands/firebase_command.dart';
 import '../lib_cli/commands/summary_command.dart';
 import '../lib_cli/commands/doctor_command.dart';
@@ -12,6 +13,7 @@ Future<void> main(List<String> args) async {
   )
     ..addCommand(SyncCommand())
     ..addCommand(ValidateCommand())
+    ..addCommand(ValidateTestspecCommand())
     ..addCommand(FirebaseCommand())
     ..addCommand(SummaryCommand())
     ..addCommand(DoctorCommand());